requirements.txt
yfinance
psycopg2-binary
pandas
python-dotenv
streamlit
sqlalchemy